+++ tests/parser/parser.test.ts
@@ -41,7 +41,7 @@ describe('Test Parser', () => {
 							cn: 'Jane Doe',
 							role: 'REQ-PARTICIPANT',
 							partstat: 'ACCEPTED',
-							delegatedFrom: 'mailto:bob@example.com',
+							delegatedFrom: 'bob@example.com',
 							address: 'ildoit@example.com',
 							scheduleStatus: [ '2.0', '2.4' ]
 						}