+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## [0.5.0] - unreleased
 
 - Added citation for regex from Ruby iCalendar library
+- Added rfc link to readme
 
 ## [0.4.0] - 2020-05-07