+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 - Added citation for regex from Ruby iCalendar library
 - Added rfc link to readme
 - Updated dtStamp as optional, current UTC time used when not set
+- Added schedule status event parameter, dquote parsing
 
 ## [0.4.0] - 2020-05-07