+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 - Added rfc link to readme
 - Updated dtStamp as optional, current UTC time used when not set
 - Added schedule status event parameter, dquote parsing
+- Updated delegatedTo/From parsing to remove 'mailto:'
 
 ## [0.4.0] - 2020-05-07